feat: Add CORS middleware and improve translation
Introduced CORS middleware to handle cross-origin requests and updated the translation system to support dynamic placeholder replacements in localized strings. This enhances the application's flexibility and internationalization capabilities.

Additionally, updated `firebase-admin` and `firebase-functions` to their latest compatible versions and adjusted the script loading in `index.html` for better compatibility with Babel. The `t` function in `App.tsx` and `useLocalization.ts` was refactored to accept and process replacements, allowing for more dynamic and context-aware translations.